Refactor working files
