+++ src/index.html
@@ -28,6 +28,5 @@
             <input type="submit" value="Submit" class="submit">
         </section>
     </main>
-    <script src="index.js"></script>
 </body>
 </html>
