+++ src/index.js
@@ -1,16 +1,20 @@
 import './style.css';
-import { refresh, submit } from './modules/listeners.js';
-import submitScore from './modules/submitScore.js';
-import loadScores from './modules/loadScores.js';
+import './modules/submitScore.js';
+import { refresh } from './modules/listeners.js';
+import fetchData from './modules/loadScores.js';
 
-loadScores();
+const scoreContainer = document.querySelector('.score-container');
+fetchData();
 
-refresh.addEventListener('click', () => {
-  loadScores();
-});
-
-submit.addEventListener('click', () => {
-  const username = document.querySelector('.username');
-  const score = document.querySelector('.score');
-  submitScore(username.value, score.value);
+refresh.addEventListener('click', async () => {
+  const scores = await fetchData();
+  scoreContainer.innerHTML = '';
+  scores.sort((a, b) => a.score - b.score).forEach((element) => {
+    const scoreData = document.createElement('tr');
+    scoreData.innerHTML = `
+    <td>${element.user}</td>
+    <td>${element.score}</td>
+    `;
+    scoreContainer.appendChild(scoreData);
+  });
 });
