+++ src/modules/submitScore.js
@@ -1,29 +1,28 @@
-import loadScores from './loadScores.js';
+import { submit } from './listeners.js';
 
-const submitScore = (username, score) => {
-  if (username || score !== '') {
-    const submitData = async () => {
-      const response = await fetch(
-        'https://us-central1-js-capstone-backend.cloudfunctions.net/api/games/kN0uWyCAtmzgUqRYuCdv/scores/',
-        {
-          method: 'POST',
-          headers: {
-            'Content-Type': 'application/json; charset=UTF-8',
-          },
-          body: JSON.stringify({
-            user: `${username}`,
-            score,
-          }),
-        },
-      );
-      return response.status;
-    };
-    submitData()
-      .then((res) => res)
-      .then(() => {
-        loadScores();
-      });
-  }
+const username = document.querySelector('.username');
+const score = document.querySelector('.score');
+
+const submitData = async () => {
+  const response = await fetch(
+    'https://us-central1-js-capstone-backend.cloudfunctions.net/api/games/kN0uWyCAtmzgUqRYuCdv/scores/',
+    {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify({
+        user: username.value,
+        score: score.value,
+      }),
+    },
+  );
+  const data = await response.json();
+  return data;
 };
 
-export default submitScore;
+submit.addEventListener('click', () => {
+  submitData();
+  username.value = '';
+  score.value = '';
+});
